Refactored project to have a core library and two binaries.
This commit removes the original binary and creates a library in its
place. It also creates separate mutagen and mutagen-agent binaries. I've
tried to keep as "standard" a structure as I can. I've put the binaries
into their own subdirectories of the src/bin subdirectory, but
apparently Cargo doesn't detect this like I'd expect, so I've had to add
some additional Cargo.toml configuration to make them work. We could
just have src/bin/mutagen.rs and src/bin/mutagen-agent.rs, but I expect
each of these programs to span multiple files. I've also disabled
documentation building for both because it's unnecessary and the mutagen
binary has the same name as the library, which Cargo doesn't support
when building documentation. I may end up just having one binary for
both the core CLI and the agent (which isn't too crazy - we already do
that for the CLI and daemon) - it all just depends on how big the binary
comes out. In any case, we'd probably still keep this subdirectory
structure.

I've added the Cargo.lock file because it will be necessary for the
binaries. If, for whatever reason, some other crate wants to include the
library as a dependency, I'm not sure how things will behave - I would
hope Cargo would just ignore it? We'll have to see. I think this is the
right thing to do for a crate with binaries in it, I think it's just
generally assumed that that's the end of the dependency chain for that
crate, but we might want to go one more level with the binary, e.g. in a
GUI in a separate crate.

I've also bumped the version to 0.2.0, which will hopefully be the point
at which we replace the Go implementation.

I've added some rudimentary version infrastructure, mostly to test
linking between the binaries and the core library. It might need some
refactoring.

I've also added some additional metadata to Cargo.toml. Some of these
keys, in particular the categories and badges information, are
documented but only supported in the nightly version of Cargo, so they
issue warnings for now, but those will go away automatically in the near
future.