Implemented basic argument parsing with clap.
The clap library provides decent 99% case CLI parsing, so I'm willing to
go with that so that I don't have to develop my own auto-complete and
Levenshtein distance implementation. I still have a few qualms about the
formatting, but I can live with it.

This commit adds all of the existing subcommands and their arguments.

The one downside is that the clap library is a bit big. I've enabled LTO
in release mode to counter this, but I think it means we'll want a
separate agent binary, so I'll likely revert the previous commit and
just use simple hand-rolled argument parsing for the agent. I think this
gives us the best possible outcome, because we want nice CLI behavior
for user-facing components, but the agent is not user-facing, so it can
stand to be a bit more barebones.

In any case, with LTO+stripping, the resulting release binary is ~1 MB
for the CLI (vs ~300 KB for the "Hello world" binary), but I'm expecting
some of that cost to be amortized over future development, because the
clap library pulls in a LOT of code.