Implemented documentation and tests for hash module.
These tests are fairly basic, as I assume the underlying implementations
are properly tested. I've computed the hash constants in the tests using
OpenSSL and b2sum.

My only concern with these tests it that perhaps we should attempt to
generate much larger blocks of data, but as this code is pretty much
only glue code between traits and algorithms, I don't think it's
necessary at this time.