Started on daemon implementation and performed some refactoring.
This commit adds support for spawning the daemon on POSIX and Windows
(though the Windows implementation is still untested). The Windows
version currently requires unstable features to set process creation
flags, but these will be stabilized in Rust 1.16 which is being released
in 2 days.

It also refactors the main command entry point and subcommand handlers
to use an error_chain-based scheme to track and print errors.

Finally, it does a bit of normalizing of different use statements, their
spacing, and their ordering.