Implemented reconciliation.
This involved a (much simpler) re-implementation of diff and the
consequent removal of the diff module.

The logic is largely the same as the Go version, though explicit pattern
matching and equality comparisons are performed instead of using the
somewhat vague equalShallow method.

Tests and documentation remain to be implemented. I think path joins can
also be optimized.

I had to change the ordered_iter crate to pull from a Git repository
(I've forked the original) because they hadn't yet made a crates.io
release with the OrderedMapIterator support on OuterJoin.