Updated AppVeyor configuration to use only GNU toolchains.
The GNU toolchain is what we're going to use for the open source
binaries we distribute, so that's all we really need to test with
AppVeyor. We have to split them up because the x86_64 toolchain passes
the --large-address-aware flag to the i686 linker, and the linker
doesn't understand this option. This is probably a bug that needs to be
reported to the Rust project.