Initial but imcomplete support for edit command.
This commit implements the ephemeral "mutagen edit" command, which
starts a synchronization session between a remote endpoint and a local
temporary directory, launches the user's editor, and waits for it to
complete before terminating this session.

There are two big blockers for finishing this, both labeled as TODOs in
edit.go. The first issue is that we need to wait for an initial
synchronization session to take place before launching the editor, and
the second is that we need to ensure a final synchronization has
completed and not left any unsynchronized local changes before
terminating. It's not enough to wait for the editor to exit since it may
save changes just before exiting and the synchronization loop may not
have picked them up. As part of this, we also need to make sure that the
session is connected and the like, but that's part of ensuring
synchronization.

There are a few other session lifecycle questions that need to be
addressed as part of this work, e.g. how long the command hangs around
if it can't sync, what happens to edit directories that don't get
cleaned up (do we do housekeeping). We have to keep the temporary edit
directories around for longer than the session lifespan, otherwise we
might inadvertently wipe some remote file. That's why we store them in
~/.mutagen and not /tmp or similar.

Note that we'll also want to vendor and license
github.com/mattn/go-shellwords if we're going to ship this.