DO NOT MERGE: Implemented prototype ignore size.
This should not be merged as it is algorithmically incorrect. The
problem is that paths being ignored can change between scans based on
their size and this would cause deletion!

I like some of the implementation ideas in this design though, and I
like the idea of having mutagen list print ignored paths.